move: PropertyAdapter class
